updating the blog

+++ blog/wp-content/plugins/feedstats-de/feedstats-de-settings.php
@@ -4,12 +4,12 @@ function feedstats_admin_option_page() {
 	global $wpdb, $wp_version;
 ?>
 <div class="wrap">
-	<h2><img src="<?php echo feedstats_get_resource_url('feedstats32.gif'); ?>" alt="" width="32" height="32" /> <?php _e('FeedStats', 'feedstats'); ?></h2>
+	<h2><?php _e('FeedStats Options', 'feedstats'); ?></h2>
 <?php
-	if ( ($_POST['action'] == 'add_index') && $_POST['feedstats_add_index'] ) {
+	if ( isset($_POST['action']) && ($_POST['action'] == 'add_index') && $_POST['feedstats_add_index'] ) {
 		
 		if ( function_exists('current_user_can') && current_user_can('edit_plugins') ) {
-			check_admin_referer($FeedStats_nonce);
+			check_admin_referer('FeedStats_nonce');
 			feedstats_genereta_tables();
 		
 			echo '<div class="updated fade"><p>' . __('Allready update the tables!', 'feedstats') . '</p></div>';
@@ -18,17 +18,18 @@ function feedstats_admin_option_page() {
 		}
 	}
 
-	if ( ($_POST['action'] == 'insert') && $_POST['fs_ifs_save'] ) {
+	if ( isset($_POST['action']) && ($_POST['action'] == 'insert') && $_POST['fs_ifs_save'] ) {
 	
 		if ( function_exists('current_user_can') && current_user_can('edit_plugins') ) {
-			check_admin_referer($FeedStats_nonce);
+			check_admin_referer('FeedStats_nonce');
 
 			// for a smaller database
 			function feedstats_get_update($option) {
-				if ( ($_POST[$option] == '0') || $_POST[$option] == '') {
+				
+				if ( ! isset($_POST[$option]) || ($_POST[$option] == '0') || $_POST[$option] == '') {
 					delete_option($option);
 				} else {
-					update_option($option , $_POST[$option]);
+					update_option($option , esc_attr( $_POST[$option] ) );
 				}
 			}
 			
@@ -46,10 +47,10 @@ function feedstats_admin_option_page() {
 		}
 	}
 	
-	if ( ($_POST['action'] == 'deactivate') && $_POST['feedstats_ifs_deactivate'] ) {
+	if ( isset($_POST['action']) && ($_POST['action'] == 'deactivate') && $_POST['feedstats_ifs_deactivate'] ) {
 
 		if ( function_exists('current_user_can') && current_user_can('edit_plugins') ) {
-			check_admin_referer($FeedStats_nonce);
+			check_admin_referer('FeedStats_nonce');
 			
 			$wpdb->query ("DROP TABLE {$wpdb->prefix}fs_data");
 			$wpdb->query ("DROP TABLE {$wpdb->prefix}fs_visits");
@@ -75,8 +76,8 @@ function feedstats_admin_option_page() {
 			<div class="postbox" >
 				<h3><?php _e('FeedStats settings', 'feedstats'); ?></h3>
 				<div class="inside">
-					<form name="form1" method="post" action="<?php echo $location; ?>">
-						<?php feedstats_nonce_field($FeedStats_nonce); ?>
+					<form name="form1" method="post" action="">
+						<?php if ( feedstats_nonce_field('FeedStats_nonce') ); ?>
 						
 						<table summary="feedstats options" class="form-table">
 							<tr valign="top">
@@ -112,11 +113,11 @@ function feedstats_admin_option_page() {
 							</tr>
 							<tr valign="top">
 								<th scope="row"><?php _e('Not tracked', 'feedstats'); ?></th>
-								<td><input name="fs_ifs_not_tracked" value="<?php echo get_option('fs_ifs_not_tracked'); ?>"  type="text" /><br /><?php _e('IP, that is supposed not to be saved, ex.: your own IP', 'feedstats'); echo '<small> ' . $_SERVER['REMOTE_ADDR'] . '</small>' ;?></td>
+								<td><input name="fs_ifs_not_tracked" value="<?php echo get_option('fs_ifs_not_tracked'); ?>" type="text" /><br /><?php _e('IP, that is supposed not to be saved, ex.: your own IP', 'feedstats'); echo '<code> ' . esc_attr( $_SERVER['REMOTE_ADDR'] ) . '</code>'; ?></td>
 							</tr>
 							<tr valign="top">
 								<th scope="row"><?php _e('Dashboardinfo', 'feedstats'); ?></th>
-								<td><input name="fs_ifs_dashboardinfo" value='1' <?php if (get_option('fs_ifs_dashboardinfo')=='1') { echo "checked='checked'";  } ?> type="checkbox" /><br /><?php _e('Statistics can be shown on the dashboard ?', 'feedstats'); ?></td>
+								<td><input name="fs_ifs_dashboardinfo" value='1' <?php if (get_option('fs_ifs_dashboardinfo') === '1') { echo "checked='checked'";  } ?> type="checkbox" /><br /><?php _e('Statistics can be shown on the dashboard ?', 'feedstats'); ?></td>
 							</tr>
 						</table>
 						<p class="submit">
@@ -129,12 +130,12 @@ function feedstats_admin_option_page() {
 		</div>
 		
 		<div id="poststuff" class="ui-sortable">
-			<div class="postbox closed" >
+			<div class="postbox" >
 				<h3><?php _e('Add Index', 'feedstats'); ?></h3>
 				<div class="inside">
 					<p><?php _e('The follow button add index to the table of thsi plugin for a better performance. Do you have install the plugin new at version 3.6.4? Then is this not necessary.', 'feedstats'); ?></p>
-					<form name="form2" method="post" action="<?php echo $location; ?>">
-						<?php feedstats_nonce_field($FeedStats_nonce); ?>
+					<form name="form2" method="post" action="">
+						<?php feedstats_nonce_field('FeedStats_nonce'); ?>
 						<p class="submit">
 							<input type="hidden" name="action" value="add_index" />
 							<input class="button" type="submit" name="feedstats_add_index" value="<?php _e('Add Index', 'feedstats'); ?> &raquo;" />
@@ -145,12 +146,12 @@ function feedstats_admin_option_page() {
 		</div>
 		
 		<div id="poststuff" class="ui-sortable">
-			<div class="postbox closed" >
+			<div class="postbox" >
 				<h3><?php _e('Delete Options', 'feedstats'); ?></h3>
 				<div class="inside">
 					<p><?php _e('The follow button delete all tables and options for the FeedStats plugin. <strong>Attention: </strong>You <strong>cannot</strong> undo any changes made by this plugin.', 'feedstats'); ?></p>
-					<form name="form2" method="post" action="<?php echo $location; ?>">
-						<?php feedstats_nonce_field($FeedStats_nonce); ?>
+					<form name="form2" method="post" action="">
+						<?php feedstats_nonce_field('FeedStats_nonce'); ?>
 						<p class="submit">
 							<input type="hidden" name="action" value="deactivate" />
 							<input class="button" type="submit" name="feedstats_ifs_deactivate" value="<?php _e('Delete Options', 'feedstats'); ?> &raquo;" />
@@ -161,7 +162,7 @@ function feedstats_admin_option_page() {
 		</div>
 		
 		<div id="poststuff" class="ui-sortable">
-			<div class="postbox closed" >
+			<div class="postbox" >
 				<h3><?php _e('Information on the plugin', 'feedstats') ?></h3>
 				<div class="inside">
 					<p><?php _e('Plugin created by <a href="http://www.anieto2k.com">Andr&eacute;s Nieto</a>, in cooperation/base with plugin <a href="http://www.deltablog.com/">PopStats</a>. German and english adjustments, little extensions and new coding by <a href="http://bueltge.de">Frank Bueltge</a>. Thx to <a href="http://blog.tomk32.de">Thomas R. Koll</a> for many improvements for a better code and performance.', 'feedstats'); ?></p>
